It almost works! But I'm gonna make a generic Generator test ctx
