+++ src/rust/generators/sysmon-generator/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod api;
+pub mod error;
+pub mod models;
