+++ src/rust/generators/sysmon-generator/src/main.rs
@@ -6,22 +6,11 @@ use opentelemetry::{
     global,
     sdk::propagation::TraceContextPropagator,
 };
-use rust_proto::graplinc::grapl::api::plugin_sdk::generators::v1beta1::{
-    server::GeneratorApi,
-    GeneratedGraph,
-    RunGeneratorRequest,
-    RunGeneratorResponse,
-};
-use sysmon_parser::SysmonEvent;
+use sysmon_generator::api;
 use tracing_subscriber::{
     prelude::*,
     EnvFilter,
 };
-
-mod error;
-mod models;
-
-use crate::error::SysmonGeneratorError;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -49,32 +38,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing::info!("logger configured successfully");
 
     let config = GeneratorServiceConfig::from_env_vars();
-    let generator = SysmonGenerator {};
+    let generator = api::SysmonGenerator {};
     server::exec_service(generator, config).await
-}
-
-pub struct SysmonGenerator {}
-
-#[async_trait::async_trait]
-impl GeneratorApi for SysmonGenerator {
-    type Error = SysmonGeneratorError;
-
-    #[tracing::instrument(skip(self, request), err)]
-    async fn run_generator(
-        &self,
-        request: RunGeneratorRequest,
-    ) -> Result<RunGeneratorResponse, Self::Error> {
-        let sysmon_event = SysmonEvent::from_str(std::str::from_utf8(&request.data)?)?;
-
-        match models::generate_graph_from_event(&sysmon_event)? {
-            Some(graph_description) => Ok(RunGeneratorResponse {
-                generated_graph: GeneratedGraph { graph_description },
-            }),
-            None => {
-                // We do not expect to handle all Sysmon event types.
-                // So we'd just return an empty Graph Description.
-                Ok(RunGeneratorResponse::default())
-            }
-        }
-    }
 }
