+++ src/rust/pipeline-ingress/src/main.rs
@@ -23,7 +23,6 @@ use rust_proto::{
     graplinc::grapl::{
         api::pipeline_ingress::v1beta1::{
             server::{
-                ConfigurationError as ServerConfigurationError,
                 PipelineIngressApi,
                 PipelineIngressServer,
             },
@@ -40,7 +39,7 @@ use rust_proto::{
     },
     protocol::{
         healthcheck::HealthcheckStatus,
-        status::Status,
+        status::Status, error::ServeError,
     },
 };
 use thiserror::Error;
@@ -116,8 +115,8 @@ enum ConfigurationError {
     #[error("failed to configure kafka client {0}")]
     Kafka(#[from] KafkaConfigurationError),
 
-    #[error("failed to configure gRPC server {0}")]
-    Server(#[from] ServerConfigurationError),
+    #[error("ServeError {0}")]
+    ServeError(#[from] ServeError),
 
     #[error("missing environment variable {0}")]
     EnvironmentVariable(#[from] VarError),
